feat(WalletProvider): refactor to make wallet connection available on whole website
