+++ frontend/alyra_solana_project/src/app/layout.tsx
@@ -1,6 +1,7 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import "./globals.css";
+import AppWalletProvider from "@/contexts/AppWalletProvider";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -16,7 +17,9 @@ export default function RootLayout({
 }>) {
 	return (
 		<html lang="en">
-			<body className="min-h-screen h-full">{children}</body>
+			<body className="min-h-screen h-full">
+				<AppWalletProvider>{children}</AppWalletProvider>
+			</body>
 		</html>
 	);
 }
