+++ frontend/alyra_solana_project/src/components/AppHeader/connectWallet.tsx
@@ -1,45 +1,15 @@
 "use client";
-import {
-	ConnectionProvider,
-	WalletProvider,
-} from "@solana/wallet-adapter-react";
-import { WalletAdapterNetwork } from "@solana/wallet-adapter-base";
-import { PhantomWalletAdapter } from "@solana/wallet-adapter-wallets";
-import {
-	WalletModalProvider,
-	WalletMultiButton,
-} from "@solana/wallet-adapter-react-ui";
-import { clusterApiUrl } from "@solana/web3.js";
-import { useMemo } from "react";
 
-// Default styles that can be overridden by your app
-import "@solana/wallet-adapter-react-ui/styles.css";
-
-function Wallet({ children }: { children: React.ReactNode }) {
-	// Can be set to 'devnet', 'testnet', or 'mainnet-beta'
-	const network = WalletAdapterNetwork.Devnet;
-
-	// You can also provide a custom RPC endpoint
-	const endpoint = useMemo(() => clusterApiUrl(network), [network]);
-
-	// @solana/wallet-adapter-wallets includes all the adapters but supports tree shaking --
-	// Only the wallets you configure here will be compiled into your application
-	const wallets = useMemo(() => [new PhantomWalletAdapter()], []);
-
-	return (
-		<ConnectionProvider endpoint={endpoint}>
-			<WalletProvider wallets={wallets} autoConnect>
-				<WalletModalProvider>{children}</WalletModalProvider>
-			</WalletProvider>
-		</ConnectionProvider>
-	);
-}
+import { WalletMultiButton } from "@solana/wallet-adapter-react-ui";
 
 export function ConnectWallet() {
 	return (
-		<Wallet>
-			{/* rgb(253 224 71 / var(--tw-bg-opacity)) /* #fde047 */}
-			<WalletMultiButton style={{ color: "black", backgroundColor: "#fde047", borderRadius: "30px", }} />
-		</Wallet>
+		<WalletMultiButton
+			style={{
+				color: "black",
+				backgroundColor: "#fde047",
+				borderRadius: "30px",
+			}}
+		/>
 	);
 }
